requirements.txt
 Django>=1.8,<2.0
 psycopg2
 djangorestframework
 django-webpack-loader==0.4.1